task 4

+++ main.py
@@ -1,4 +1,4 @@
-from PyQt5.QtWidgets import QApplication, QMainWindow, QLabel
+from PyQt5.QtWidgets import QApplication, QMainWindow, QLabel, QComboBox, QPushButton
 from PyQt5.QtGui import QPixmap
 from PyQt5.Qt import Qt
 import requests
@@ -10,32 +10,41 @@ SCREEN_SIZE = [600, 450]
 
 class StaticMapAPI:
     map_file = 'map.png'
+    map_file_for_sat = 'map.jpg'
     _server = 'https://static-maps.yandex.ru/1.x/'
     ll = [37.620070, 55.753630]
     size = [600, 450]
     l = 'map'
-    z = 13
-    ZOOM_MAX = 20
-    ZOOM_MIN = 1
+    spn = 0.002
+    MAX_SPN = 65.536
+    MIN_SPN = 0.001
     pixmap = None
 
     def set_params(self, **params):
-        pass
+        self.l = params.get('l', self.l)
+        self.spn = params.get('spn', self.spn)
+        self.size = params.get('size', self.size)
+        self.ll = params.get('ll', self.ll)
 
     def update_map(self):
         params = {
             'll': f'{self.ll[0]},{self.ll[1]}',
             'size': f'{self.size[0]},{self.size[1]}',
             'l': self.l,
-            'z': self.z
+            'spn': f'{self.spn},{self.spn}'
         }
         response = requests.get(self._server, params=params)
         if not response:
             raise Exception(response.reason)
 
-        with open(self.map_file, "wb") as file:
-            file.write(response.content)
-        self.__pixmap = QPixmap(self.map_file)
+        if self.l == 'sat':
+            with open(self.map_file_for_sat, "wb") as file:
+                file.write(response.content)
+            self.__pixmap = QPixmap(self.map_file_for_sat)
+        else:
+            with open(self.map_file, "wb") as file:
+                file.write(response.content)
+            self.__pixmap = QPixmap(self.map_file)
 
     def get_pixmap(self, update=False):
         if update:
@@ -45,19 +54,19 @@ class StaticMapAPI:
     def get_size(self):
         return self.size
 
-    def zoom_in(self, step):
-        self.z += step
-        if self.z > self.ZOOM_MAX:
-            self.z = self.ZOOM_MAX
+    def spn_in(self):
+        self.spn /= 2
+        if self.spn < self.MIN_SPN:
+            self.spn = self.MIN_SPN
 
-    def zoom_out(self, step):
-        self.z -= step
-        if self.z < self.ZOOM_MIN:
-            self.z = self.ZOOM_MIN
+    def spn_out(self):
+        self.spn *= 2
+        if self.spn > self.MAX_SPN:
+            self.spn = self.MAX_SPN
 
     def move(self, x=0, y=0):
-        self.ll[0] += x * 1.05 ** self.z * 0.05
-        self.ll[1] += y * 1.05 ** self.z * 0.05
+        self.ll[0] += x * self.spn * 1 * 2.5
+        self.ll[1] += y * self.spn * (85 / 180) * 2.5
 
 
 class MapsAPI(QMainWindow):
@@ -66,21 +75,30 @@ class MapsAPI(QMainWindow):
         self.static_map_api = StaticMapAPI()
         self.initUI()
 
+        self.btn_update_map.clicked.connect(self.update_map)
+        # self.mode.currentTextChanged.connect(self.update_map)
+
     def initUI(self):
         self.setWindowTitle('Maps API')
         self.setFixedSize(*SCREEN_SIZE)
         self.label_map = QLabel(self)
+        self.label_map.move(100, 0)
+        self.mode = QComboBox(self)
+        self.mode.addItems(['map', 'sat', 'skl'])
+        self.btn_update_map = QPushButton('Обновить', self)
+        self.btn_update_map.move(0, 100)
         self.update_map()
 
     def update_map(self):
+        self.static_map_api.set_params(l=self.mode.currentText())
         self.label_map.setPixmap(self.static_map_api.get_pixmap(update=True))
         self.label_map.resize(*self.static_map_api.get_size())
 
     def keyPressEvent(self, event):
         if event.key() == Qt.Key_PageUp:
-            self.static_map_api.zoom_in(1)
+            self.static_map_api.spn_in()
         elif event.key() == Qt.Key_PageDown:
-            self.static_map_api.zoom_out(1)
+            self.static_map_api.spn_out()
         elif event.key() == Qt.Key_Up:
             self.static_map_api.move(x=0, y=1)
         elif event.key() == Qt.Key_Down:
@@ -94,7 +112,10 @@ class MapsAPI(QMainWindow):
         self.update_map()
 
     def closeEvent(self, *args, **kwargs):
-        os.remove(self.static_map_api.map_file)
+        if self.static_map_api.map_file in os.listdir():
+            os.remove(self.static_map_api.map_file)
+        if self.static_map_api.map_file_for_sat in os.listdir():
+            os.remove(self.static_map_api.map_file_for_sat)
 
 
 if __name__ == '__main__':
